requirements.txt
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0

